streamlit_app/requirements.txt
streamlit>=1.28.0
pyyaml>=6.0
confluent-kafka>=2.3.0
